CHG: OperandSize byte has been extended with a OperandFamily bits
The OperandFamily defines the target (or family) of the op-code. There are currently 3 families/targets defined. Integer, Control and Float. The integer is the normal target - all operations work with the integer ALU. The Control is for MMU, CPU, FPU control instructions. For instance if family is 'CONTROL' you can move to the control-registers. Not all instructions are supported. In general only move instructions are supported.