CHG: Compile unit now handles import/exports explicitly
This was previously done by directly looking up the affected export and adding us to the list of resolve points. Instead we create a local resolve point in the compile unit when merging the units and recomputes the address before adding it to the export resolve point list.

The linker now has proper absolute addresses and can replace directly.

The merge/calculation step can also be done in the linker - this is subject to change.