CHG: Branching flag added as feature to all branching instructions
I need this flag. Since handling of reading data from a label differs.

call <label>  ; jump to the absolute address of label
move d0, label  ; read the value out from the absolute address of label

The behaviour in the CPU requires a 'flag' on the instruction if the memory should be fetched (during decoding) or not. This must be decided when emitting the op-code and the tagging is done.