FIX: Introduced an operand description flag to distinguish how instr. encoding should be for memory reads
Currently the name 'Branching' is wrong. But it is used to describe if the memory access should happen at the aforementioned location or if we aim to use the access..

The difference is:
move.l d0, label     <- this should READ the data from the label (i.e. memory read occurs)
call label    <- this should NOT read label but put the address of label (i.e. no memory read occurs)

It was first implemented for branching, but later also discovered that 'lea' requires this flag to be set.