CHG: Offset calculation is 'minus 1'.
I need to verify if this is stays the same regardless of how relative branches are made. I am quite sure it depends on the op.size (i.e. -1 for bytes, -2 for words and -4 for dwords).

The reason for this is that the offset is saved relative to the position of the offset-location in the instruction and the jump is done relative to the start of next instruction. Therefore the offset (in case of bytes) must be subtracted with -1.