FIX: Moved OpCodes around
The 0xf0-0xfe range (RET, NOP, RTI) is not in the 0x60 range.

The 0x60 range is moved to 0xf0-0xfe range. These are reserved for instruction set extensions.

Reason for the move is simpler HW logic to select correct instr. decoder. This 0xf- becomes very easy to test in HW.

The main idea behind this is to allow instruction set extensions to be added. For instance FPU, SIMD, etc..