CHG: Added comment about what I think might be a good idea
Not thought this through yet. But I believe it might be a good idea to reserve some space for ExceptionControl block and pass a pointer to it for the exception handler to pick up.

This way a debugger can trap the exception handler - inspect - and then allow the CPU to continue.

Also should probably consider if this a valid 'trap' scenario for basic debugging.