CHG: Timer interface changed
Now takes a pointer to a timer-configuration memory block. When the CPU is started with a proper amount of memory. A chunk of that memory is reserved for various functions. The timer is now properly mapped to that memory. So manipulating it from code will work.

The timer it-self runs in a thread.