WIP: Lot's of work on the SIMD part and how extensions should work
Extensions need much more 'thought' in general. Basic instr. redirection now works. But the whole code chain is very much patched together with tape..  This won't work..

Need a more distinct way to add extensions and INTEGRATE them with the flow path..

This also has a great effect on the assembler and the way it needs to work. I would like a minimal effort in terms of adding new extension or changing existing ones.