CHG: Now there is a solution where extension decoders are created on demand and follows the main decoder
This allows extensions to follow their cores. As the base-decoders are core and pipeline specific. However, it doesn't solve different core's for extensions.

I.e. extensions can be scheduled on different cores..

No sure how to solve this if the instructions are interleaved..  One way would be to 'block' the core that is running and let it wait - the CORE instr. pointer would have to be 'moved' afterwards - ergo, compuation of instr.size must be shared and/or returned.