WIP: Refactoring and decoupling of instruction-set definition/decoder/execution
The instruction set decoder and execution modules DO NOT talk to each other. The decoder now translates the binary into something else and then passes it on. The CPU is pushing it on a dispatcher queue which is then calling the execution module.

We can also schedule various instruction-set's differently in the same dispatcher. So it's agnostic in that sense.

The dispatcher buffer can be dynamically set.

Note: That decoder output is quite 'large' (i.e. takes many bytes) - there is quite a bit of redundant information and debugging data in there.